+++ packages/nextra/src/server/mdx-plugins/remark-headings.ts
@@ -2,8 +2,8 @@ import Slugger from 'github-slugger'
 import type { Parent, Root } from 'mdast'
 import type { Plugin } from 'unified'
 import { visit } from 'unist-util-visit'
-import { MARKDOWN_EXTENSION_REGEX } from '../constants'
-import type { Heading } from '../types'
+import { MARKDOWN_EXTENSION_REGEX } from '../../constants'
+import type { Heading } from '../../types'
 import type { HProperties } from './remark-custom-heading-id'
 
 const getFlattenedValue = (node: Parent): string =>
