+++ packages/nextra/src/server/mdx-plugins/remark-link-rewrite.ts
@@ -1,7 +1,7 @@
 import type { Root } from 'mdast'
 import type { Plugin } from 'unified'
 import { visit } from 'unist-util-visit'
-import { EXTERNAL_URL_REGEX } from '../constants'
+import { EXTERNAL_URL_REGEX } from '../../constants'
 
 export type RemarkLinkRewriteOptions = {
   pattern: RegExp
