+++ packages/nextra/src/server/mdx-plugins/remark-mdx-disable-explicit-jsx.ts
