+++ packages/nextra/src/server/theme.json
