+++ packages/nextra/src/server/webpack-plugins/nextra.ts
@@ -1,10 +1,10 @@
 import { join, relative } from 'node:path'
 import pkg from 'graceful-fs'
 import type { Compiler } from 'webpack'
-import { CHUNKS_DIR, CWD } from '../constants'
+import { CHUNKS_DIR, CWD } from '../../constants'
 import { PAGES_DIR } from '../file-system'
-import { collectPageMap } from '../server/page-map'
-import { logger } from '../server/utils'
+import { collectPageMap } from '../page-map'
+import { logger } from '../utils'
 
 const fs = pkg.promises
